Rename plugin: register => registry
