+++ plugins/registry.js
